add code to parse strike discuz code

+++ BUClient/Classes/Models/BUCModels.h
@@ -55,7 +55,7 @@
 @end
 
 
-typedef NS_OPTIONS(uint32_t, BUCLinkType) {
+typedef NS_OPTIONS(uint8_t, BUCLinkType) {
     BUCPostLink = 1 << 0,
     BUCMailLink = 1 << 1,
     BUCUrlLink = 1 << 2
